chỉnh sửa chút trong detail với favoriteproduct

+++ src/app/client/profile/fovoriteproduct/fovoriteproduct.component.ts
@@ -46,13 +46,13 @@ export class FovoriteproductComponent implements OnInit {
     );
   }
   remove(FP_id){
-    console.log(FP_id);
-    this.productservice.removeFavoriteProduct(FP_id).subscribe(
-      res=>{
-        console.log(res);
-        this.showFavoriteProduct();
-      }
-    );
+    if(confirm("Bạn có muốn xóa sản phẩm ra khỏi danh sách sản phẩm yêu thích")){
+      this.productservice.removeFavoriteProduct(FP_id).subscribe(
+        res=>{
+          console.log(res);
+          this.showFavoriteProduct();
+        }
+      );
+    }
   }
-
 }
